docs: add tip on swapping tx/rx pins
First of all thanks for creating this project, it works quite well! Recently I've set up 2 ACs with this integration and couldn't get them to work initially because the AC's board internal wiring doesn't reflect the diagram (I think the TX and RX connections are swapped). The docs mention this but are not explicit about the options there are (physical and software changes) to solve the problem so I figured I could add them here to help others facing the same problem.